feat!: simplify struct interfaces
We were using Cow as a holdover from previous clients. In those, there was some benefit of being able to use static strings. However, in WIT, we always have to go to a string eventually, and then we might as well make them allocated earlier on and remove a bunch of awkward usage patterns.